Update OpenSans font and enhance planogram/cluster features
Updated OpenSans.ttf with new glyphs and font table changes to improve appearance and compatibility. Added and updated front-end TypeScript models, services, Vue components, and Pinia stores for planogram and cluster management. Introduced comprehensive SCSS styles and custom fonts for a modern UI. Enhanced JointJS+ shape definitions and SVG icons. Extended C# backend controllers, services, repositories, DTOs, and database models to support new cluster/layout flows, flexible querying, and improved error handling. Updated submodules to reflect shared library changes. No application logic was changed in the font update.

+++ Entities/ClusterAggregate/ClusterPart.cs
@@ -23,7 +23,7 @@ public partial class ClusterPart : BaseEntity<long>, IAggregateRoot
 
     public DateTime? DateUpdated { get; set; }
 
-    public int? ScratchPadId { get; set; }
+    public long? ScratchPadId { get; set; }
 
     public virtual Cluster Cluster { get; set; } = null!;
